+++ core/src/main/java/no/nav/saf/anticorruptionlayer/joark/hentjournalsakinfo/rjoark900/DokumentInfoDto.java
@@ -23,5 +23,6 @@ public class DokumentInfoDto {
 	private String tittel;
 	private SkjermingTypeCode skjerming;
 	private List<VariantDto> varianter;
+	private Long origJournalpostId;
 	private List<LogiskVedleggDto> logiske;
 }