+++ core/src/main/java/no/nav/saf/domain/visningsmodell/Tilleggsopplysning.java
@@ -1,0 +1,12 @@
+package no.nav.saf.domain.visningsmodell;
+
+import lombok.Value;
+
+/**
+ * @author Joakim Bjørnstad, Jbit AS
+ */
+@Value
+public class Tilleggsopplysning {
+	private final String nokkel;
+	private final String verdi;
+}